+++ cells/cell.go
@@ -20,7 +20,6 @@ import (
 	"github.com/tideland/golib/identifier"
 	"github.com/tideland/golib/logger"
 	"github.com/tideland/golib/loop"
-	"github.com/tideland/golib/monitoring"
 	"github.com/tideland/golib/scene"
 )
 
@@ -210,8 +209,10 @@ func (c *cell) stop() error {
 
 // backendLoop is the backend for the processing of messages.
 func (c *cell) backendLoop(l loop.Loop) error {
-	monitoring.IncrVariable(identifier.Identifier("cells", c.env.ID(), "total-cells"))
-	defer monitoring.DecrVariable(identifier.Identifier("cells", c.env.ID(), "total-cells"))
+	totalCellsID := identifier.Identifier("cells", c.env.ID(), "total-cells")
+	c.env.monitoring().IncrVariable(totalCellsID)
+	defer c.env.monitoring().DecrVariable(totalCellsID)
+
 	for {
 		select {
 		case <-l.ShallStop():
@@ -220,9 +221,9 @@ func (c *cell) backendLoop(l loop.Loop) error {
 			if event == nil {
 				panic("received illegal nil event!")
 			}
-			measuring := monitoring.BeginMeasuring(c.measuringID)
-			defer measuring.EndMeasuring()
+			measuring := c.env.monitoring().BeginMeasuring(c.measuringID)
 			err := c.behavior.ProcessEvent(event)
+			measuring.End()
 			if err != nil {
 				logger.Errorf("cell %q processed event %q with error: %v", c.id, event.Topic(), err)
 				return err