+++ cells/doc.go
@@ -77,7 +77,7 @@ import (
 
 // PackageVersion returns the version of the version package.
 func PackageVersion() version.Version {
-	return version.New(4, 2, 2)
+	return version.New(4, 3, 0)
 }
 
 // EOF